+++ src/js/page/odoo/commands/common/exportfile.mjs
@@ -8,16 +8,19 @@ import save2file from '@terminal/utils/save2file';
 import Recordset from '@terminal/core/recordset';
 import csvStringify from '@terminal/utils/csv';
 import replacer from '@terminal/utils/stringify_replacer';
+import createZip from '@terminal/utils/zip';
 import xmlStringify from '@odoo/net_utils/xml';
 import uniqueId from '@trash/utils/unique_id';
 import {ARG} from '@trash/constants';
 import type {CMDCallbackArgs, CMDCallbackContext, CMDDef} from '@trash/interpreter';
 import type Terminal from '@terminal/terminal';
 
+const RECORDSET_FORMATS = ['csv', 'xml', 'zip'];
+
 async function cmdExportFile(this: Terminal, kwargs: CMDCallbackArgs, ctx: CMDCallbackContext): Promise<string> {
   const filename = kwargs.filename ? kwargs.filename : `${uniqueId('term')}_${new Date().getTime()}.${kwargs.format}`;
   let mime = '';
-  let data = '';
+  let data: mixed = '';
   const is_recordset = kwargs.value instanceof Recordset;
   if (kwargs.format === 'json') {
     mime = 'text/json';
@@ -26,7 +29,7 @@ async function cmdExportFile(this: Terminal, kwargs: CMDCallbackArgs, ctx: CMDCa
     } else {
       data = JSON.stringify(kwargs.value, replacer, 4);
     }
-  } else if (kwargs.format === 'csv' || kwargs.format === 'xml') {
+  } else if (RECORDSET_FORMATS.includes(kwargs.format)) {
     if (!is_recordset) {
       throw new Error(i18n.t('cmdExportFile.invalidValue', 'Invalid value: must be a recordset with csv and xml'));
     }
@@ -36,6 +39,18 @@ async function cmdExportFile(this: Terminal, kwargs: CMDCallbackArgs, ctx: CMDCa
     } else if (kwargs.format === 'xml') {
       mime = 'text/xml';
       data = await xmlStringify(kwargs.value, await this.getContext());
+    } else if (kwargs.format === 'zip') {
+      const field_names = kwargs.value.fieldNames;
+      if (!field_names.includes(kwargs.field_name)) {
+        throw new Error(i18n.t('cmdExportFile.invalidFieldName', "Invalid field name: no field named ‘{{fieldName}}’ is found to represent the name", {fieldName: kwargs.field_name}));
+      }
+      if (!field_names.includes(kwargs.field_data)) {
+        throw new Error(i18n.t('cmdExportFile.invalidFieldData', "Invalid field name: no field named ‘{{fieldData}}’ is found to represent the data", {fieldData: kwargs.field_data}));
+      }
+
+      mime = 'application/zip';
+      const zip_values = kwargs.value.toJSON().filter(rec => rec[kwargs.field_name] && rec[kwargs.field_data]).map(rec => [rec[kwargs.field_name], rec[kwargs.field_data], {base64: true}]);
+      data = await createZip(zip_values);
     }
   } else if (kwargs.format === 'raw') {
     mime = 'application/octet-stream';
@@ -55,9 +70,11 @@ export default function (): Partial<CMDDef> {
     detail: i18n.t('cmdExportFile.detail', 'Exports the command result to a text/json file.'),
     args: [
       [ARG.Flag, ['no-header', 'no-header'], false, i18n.t('cmdExportFile.args.noHeader', "Don't use header"), false],
-      [ARG.String, ['f', 'format'], false, i18n.t('cmdExportFile.args.format', 'The format to use for exporting'), 'json', ['json', 'csv', 'xml', 'raw']],
+      [ARG.String, ['f', 'format'], false, i18n.t('cmdExportFile.args.format', 'The format to use for exporting'), 'json', ['json', 'csv', 'xml', 'zip', 'raw']],
       [ARG.String, ['fn', 'filename'], false, i18n.t('cmdExportFile.args.filename', 'The filename')],
       [ARG.String, ['d', 'delimiter'], false, i18n.t('cmdExportFile.args.delimiter', 'The delimiter'), ','],
+      [ARG.String, ['fname', 'field-name'], false, i18n.t('cmdExportFile.args.fieldName', 'The field representing the file name'), 'filename'],
+      [ARG.String, ['fdata', 'field-data'], false, i18n.t('cmdExportFile.args.fieldData', 'The field representing the file data'), 'datas'],
       [ARG.Any, ['v', 'value'], true, i18n.t('cmdExportFile.args.value', 'The value to export')],
     ],
     example: "-c 'search res.partner'",