+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 **11.11.0**
 
 ```
-FIX: Command 'exportfile': Wrong object (issue #147)
+ADD: Command 'zip': Create ZIP object file
+ADD: Internal function 'encode': Encode data (base64)
+ADD: Internal function 'decode': Decode data (base64)
+
+IMP: Command 'exportfile': New format 'raw' to generic binary
+
+FIX: Command 'exportfile': Incorrect handling of recordset (issue #147)
 ```
 
 **11.10.1**